# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CandleStick {
  date: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}

export type IndicatorsPattern =
  (typeof IndicatorsPattern)[keyof typeof IndicatorsPattern];

export const IndicatorsPattern = {
  Breakout: "Breakout",
  Oversold: "Oversold",
  Overbought: "Overbought",
  Trend_Reversal: "Trend Reversal",
  Neutral: "Neutral",
} as const;

export type IndicatorsSentiment =
  (typeof IndicatorsSentiment)[keyof typeof IndicatorsSentiment];

export const IndicatorsSentiment = {
  Bullish: "Bullish",
  Bearish: "Bearish",
  Neutral: "Neutral",
} as const;

export interface Indicators {
  rsi?: number | null;
  macd?: number | null;
  macdSignal?: number | null;
  macdHistogram?: number | null;
  bbUpper?: number | null;
  bbMiddle?: number | null;
  bbLower?: number | null;
  currentPrice: number;
  priceChange: number;
  priceChangePct: number;
  pattern: IndicatorsPattern;
  winRate: number;
  expectedMove: number;
  sentiment: IndicatorsSentiment;
}

export interface StockDataResponse {
  ticker: string;
  companyName: string;
  candles: CandleStick[];
  indicators: Indicators;
}

export interface AnalysisRequest {
  rsi?: number | null;
  macd?: number | null;
  macdSignal?: number | null;
  currentPrice: number;
  pattern: string;
  sentiment: string;
  winRate: number;
  expectedMove: number;
  companyName: string;
}

export type AnalysisResponseSentiment =
  (typeof AnalysisResponseSentiment)[keyof typeof AnalysisResponseSentiment];

export const AnalysisResponseSentiment = {
  Bullish: "Bullish",
  Bearish: "Bearish",
  Neutral: "Neutral",
} as const;

export interface AnalysisResponse {
  analysis: string;
  sentiment: AnalysisResponseSentiment;
  actionLabel: string;
}

export interface ErrorResponse {
  error: string;
  suggestions?: string[];
}