# lib/api-zod/src/generated/types/analysisRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface AnalysisRequest {
  rsi?: number | null;
  macd?: number | null;
  macdSignal?: number | null;
  currentPrice: number;
  pattern: string;
  sentiment: string;
  winRate: number;
  expectedMove: number;
  companyName: string;
}