# lib/api-zod/src/generated/types/analysisResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResponseSentiment } from "./analysisResponseSentiment";

export interface AnalysisResponse {
  analysis: string;
  sentiment: AnalysisResponseSentiment;
  actionLabel: string;
}