# lib/api-zod/src/generated/types/candleStick.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface CandleStick {
  date: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
}