# lib/api-zod/src/generated/types/indicators.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { IndicatorsPattern } from "./indicatorsPattern";
import type { IndicatorsSentiment } from "./indicatorsSentiment";

export interface Indicators {
  rsi?: number | null;
  macd?: number | null;
  macdSignal?: number | null;
  macdHistogram?: number | null;
  bbUpper?: number | null;
  bbMiddle?: number | null;
  bbLower?: number | null;
  currentPrice: number;
  priceChange: number;
  priceChangePct: number;
  pattern: IndicatorsPattern;
  winRate: number;
  expectedMove: number;
  sentiment: IndicatorsSentiment;
}