# lib/api-zod/src/generated/types/stockDataResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ET RetailX Chart Pattern Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { CandleStick } from "./candleStick";
import type { Indicators } from "./indicators";

export interface StockDataResponse {
  ticker: string;
  companyName: string;
  candles: CandleStick[];
  indicators: Indicators;
}